Move PersonType to business.person package
