+++ src/ch/team2/business/person/IPerson.java
@@ -1,7 +1,5 @@
 package ch.team2.business.person;
 
-import ch.team2.business.PersonType;
-
 /**
  * Interface for PersonBase
  */
