+++ src/ch/team2/business/person/PersonFactory.java
@@ -1,6 +1,5 @@
 package ch.team2.business.person;
 
-import ch.team2.business.PersonType;
 import ch.team2.persistence.person.IPersonDAO;
 import ch.team2.persistence.person.IPersonDAOFactory;
 import ch.team2.persistence.person.PersonDAOFactory;
